Move dormant notifications to notifications array
